No AuthenticationProvider found for org.springframework.security.authentication.UsernamePasswordAuthenticationToken Username

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
         return "index";
     }
     @GetMapping("/{id}")
-    public String getUserById(@PathVariable("id") long id, Model model) {
+    public String getUserById(@PathVariable("id") int id, Model model) {
         model.addAttribute("user", userService.findById(id));
         return "user";
     }
